Data.Atomics.Internal

Description

This module provides only the raw primops (and necessary types) for atomic operations.

Documentation

casIntArray# :: MutableByteArray# d -> Int# -> Int# -> Int# -> State# d -> (#TupleRep ([] RuntimeRep), IntRep, State# d, Int##)

Given an array, an offset in Int units, the expected old value, and the new value, perform an atomic compare and swap i.e. write the new value if the current value matches the provided old value. Returns the value of the element before the operation. Implies a full memory barrier.

fetchAddIntArray# :: MutableByteArray# d -> Int# -> Int# -> State# d -> (#TupleRep ([] RuntimeRep), IntRep, State# d, Int##)

Given an array, and offset in Int units, and a value to add, atomically add the value to the element. Returns the value of the element before the operation. Implies a full memory barrier.

readForCAS# :: MutVar# RealWorld a -> State# RealWorld -> (#State# RealWorld, Ticket a#)

casMutVarTicketed# :: MutVar# RealWorld a -> Ticket a -> Ticket a -> State# RealWorld -> (#State# RealWorld, Int#, Ticket a#)

casArrayTicketed# :: MutableArray# RealWorld a -> Int# -> Ticket a -> Ticket a -> State# RealWorld -> (#State# RealWorld, Int#, Ticket a#)

Unsafe, machine-level atomic compare and swap on an element within an Array.

data Ticket a

When performing compare-and-swaps, the ticket encapsulates proof that a thread observed a specific previous value of a mutable variable. It is provided in lieu of the "old" value to compare-and-swap.

Design note: Tickets exist to hide objects from the GHC compiler, which can normally perform many optimizations that change pointer equality. A Ticket, on the other hand, is a first-class object that can be handled by the user, but will not have its pointer identity changed by compiler optimizations (but will of course, change addresses during garbage collection).

Very unsafe, not to be used

ptrEq :: a -> a -> Bool